refactor: model field types
