+++ src/customTypeModelFields/separator.ts
@@ -1,0 +1,11 @@
+import type { CustomTypeModelFieldType } from "./types";
+
+/**
+ * @deprecated - Legacy field. Do not use.
+ */
+export interface CustomTypeModelSeparatorField {
+	type: typeof CustomTypeModelFieldType.Separator;
+	config?: {
+		label?: string | null;
+	};
+}
