+++ src/fields/boolean.ts
@@ -1,0 +1,6 @@
+/**
+ * A Boolean field.
+ *
+ * @see More details: {@link https://prismic.io/docs/core-concepts/boolean}
+ */
+export type BooleanField = boolean;
